Move `QuickOpenPanel` to `Overlays` module and rename it to `OverlayPanel` (it can be used by any other overlay). Make `OverlayPanel` close on `windowDidResignKey` as Xcode does
